refactor: 주기 칩 컴포넌트 코드 개선
