+++ src/page/todo/myTodo/component/CycleChip/CycleChip.tsx
@@ -1,5 +1,13 @@
 import { chipBase, selectorChip, displayChip } from './CycleChip.css';
-import type { CycleChipProps } from './CycleChip.types';
+
+export type CycleType = '매일' | '매주' | '한번';
+
+export interface CycleChipProps {
+  type: 'selector' | 'display';
+  value: CycleType;
+  selected?: boolean;
+  onClick?: (cycle: CycleType) => void;
+}
 
 const CycleChip = ({ type, value, selected, onClick }: CycleChipProps) => {
   return type === 'selector' ? (
@@ -7,7 +15,6 @@ const CycleChip = ({ type, value, selected, onClick }: CycleChipProps) => {
       type="button"
       className={`${chipBase} ${selected ? selectorChip.selected : selectorChip.deselected}`}
       onClick={() => onClick?.(value)}
-      tabIndex={0}
     >
       {value}
     </button>
