--- src/page/todo/myTodo/component/CycleChip/CycleChip.types.ts
@@ -1,8 +1,0 @@
-export type CycleType = '매일' | '매주' | '한번';
-
-export interface CycleChipProps {
-  type: 'selector' | 'display';
-  value: CycleType;
-  selected?: boolean;
-  onClick?: (cycle: CycleType) => void;
-}
